Migrate to c# 8

+++ Infostructure/SearchParameters.cs
@@ -1,15 +1,13 @@
-using System;
-
 namespace RayTracing2._0
 {
     public class SearchParameters
     {
         public Ray Ray { get; private set; }
 
-        public double MinRayCoefficient{ get; private set; }
-        public double MaxRayCoefficient{ get; private set; }
+        public float MinRayCoefficient { get; private set; }
+        public float MaxRayCoefficient { get; private set; }
 
-        public SearchParameters(Ray ray, double minRayCoefficient = 0.001, double maxRayCoefficient = double.PositiveInfinity)
+        public SearchParameters(Ray ray, float minRayCoefficient = 0.001f, float maxRayCoefficient = float.PositiveInfinity)
         {
             Ray = ray;
             MinRayCoefficient = minRayCoefficient;
