+++ Lights/DirectionLight.cs
@@ -1,8 +1,8 @@
-using System.Collections.Generic;
+using System.Numerics;
 
 namespace RayTracing2._0
 {
-    public class DirectionLight: Light
+    public class DirectionLight : Light
     {
         public Vector3 Direction;
         public DirectionLight(Vector3 direction, VecColor color, double intensity) : base(color, intensity)
@@ -11,13 +11,13 @@ namespace RayTracing2._0
             Color = color;
             Intensity = intensity;
         }
-        
-        
+
+
 
         public override SearchParameters GetSearchParametersForEclipsingObjects(Vector3 crossPoint)
         {
-            return new SearchParameters(new Ray(crossPoint, Direction), 0.001,
-                double.PositiveInfinity);
+            return new SearchParameters(new Ray(crossPoint, Direction), 0.001f,
+                float.PositiveInfinity);
         }
 
         public override Vector3 GetDirection(Vector3 point)
