+++ SceneObjects/ISceneObject.cs
@@ -1,12 +1,13 @@
-using System.Collections.Generic;
 using RayTracing2._0.Material;
+using System.Collections.Generic;
+using System.Numerics;
 
 namespace RayTracing2._0
 {
     public interface ISceneObject
     {
-        IEnumerable<(double, Vector3)> FindIntersectedRayCoefficients(Ray ray);
-        
+        IEnumerable<(float, Vector3)> FindIntersectedRayCoefficients(Ray ray);
+
         IMaterial Material { get; }
         Vector3 Location { get; }
     }
